% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve.R
\name{curve.gfamodel}
\alias{curve.gfamodel}
\title{Plot typical response curves}
\usage{
curve.gfamodel(mod = list("ths", "thc", "thn", "blt", "war", "htj", "grg",
  "grf"), cex.legend = 0.8, log = "xy")
}
\arguments{
\item{mod}{character or list of characters. A keyword or list of keywords from the list of known analytical model: "ths","thc", "thn", "blt", "war","htj". Can also be a gfamodel class object.}

\item{cex.legend}{numeric. factor for sizing the legend.}

\item{log}{character string which contains "x" if the x axis is to be logarithmic, "y" if the y axis is to be logarithmic and "xy" or "yx" if both axes are to be logarithmic.}
}
\description{
Plot one or several typical response to pumping curve(s) for known groundwater flow analytical models.
}
\details{
If \verb{mod} is a gfamodel class object, only the model type of the \verb{mod} object is used for plot. If available in the \verb{mod} object, analytical solution parameters are ignored.
}
\seealso{
\code{\link{gfamodel}}
}
