% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawdowns-class.R, R/show.R
\docType{class}
\name{drawdowns-class}
\alias{drawdowns-class}
\alias{show,drawdowns-method}
\title{Class "drawdowns"}
\usage{
\S4method{show}{drawdowns}(object)
}
\arguments{
\item{object}{A drawdowns class object}
}
\description{
Class for representing drawdowns (and corresponding derivative of the drwadowns) at a well or a piezometer, for positive times during or after pumping (recovery).
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Display the object, by printing.
}}

\section{Slots}{

 \describe{
   \item{\code{dd}:}{Object of class \code{"data.frame"}, containing the times and the corresponding numerical values of the drawdowns.}
   \item{\code{ldiff}:}{Object of class \code{"data.frame"}, containing the times and the correspondinf davlues of the calculated derivative of the drawdowns.}
 }
}

\author{
François Bertone
}
