% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawdowns.R
\name{drawdowns}
\alias{drawdowns}
\title{Create drawdowns objects}
\usage{
drawdowns(t, s = NULL, q = NULL, fn = NULL, model = NULL,
  sample = FALSE, miscDdOptions = list())
}
\arguments{
\item{t}{numeric vector or matrix; time from the begining of the pumping test unless Argarwal type is selected. In this case, time from the begining of the recovery period.}

\item{s}{numeric vector representing the drawdown measured for the corresponding time. If s is null and t a matrix, s is set to the second column of the t matrix. If s is NULL and t a vector, s is calculated from the gfamodel parameter model.}

\item{q}{numeric or numeric matrix; ignored if fn = NULL, otherwise, q is a parameter giving the pumping rate. In case of constant pumping rate, value of the pumping rate. In case of variable pumping rate, a double columns matrix, each row containg time (since the begining of the test) at which the period ends and flow rate during the period.}

\item{fn}{a character string setting the nature of the treatment applied to the original t and s values. See details.}

\item{model}{object of class gfamodel used to calculate the s value for every t data. If s is not null, model is ignored.}

\item{sample}{logical. If TRUE, data set is sampled at regular intervals on a logarithmic scale (points are extracted from the data set). A number of samples can be indicated in the list of parameters with the nval keyword.}

\item{miscDdOptions}{list with named arguments that provide additional control over the treatment applied to the original t and s values. For example: list(defunc = "std"). If the list is empty, drawdowns uses default values. The following parameters can be set:
\describe{
    \item{nval}{maximum number of resulting data points when sampling raw data is activated.}
    \item{defunc}{nature of the function used for calculating the derivative. See Details.}
    \item{df}{additional parameter for the function used for calculating the derivative. For spline function, represents the degree of freedom. See Details.}
    \item{filter.width}{a positive integer defining the window width used for robust regression filter of the derivative values. Default value is 7. See \code{\link{robreg.filter}}.}
    \item{filter.method}{a character string containing the method to be used for robust regression filter of the derivative values. Possible options are: "DR" for Deepest Regression, "LMS" for Least Median of Squares regression, "LQD" for Least Quartile Difference regression, "LTS" for Least Trimmed Squares regression, "MED" for Median, "RM" for Repeated Median regression. Default method is "RM". See \code{\link{robreg.filter}}.}
    }}
}
\value{
An object of class drawdowns containing time or reduced time and drawdowns or reduced drawdowns ready for adjustement with a gfamodel
}
\description{
Function that creates drawdowns objects, i.e. objects that hold the information necessary for being adjusted to an analytic groundwater flow model (class gfamodel) for the interpretation of a constant rate or variable rate pumping test. It fundamentally contains time (or reduced time) and drawdowns (or reduced drawdowns). It can be constructed from variable flow rate test and/or for recovery period.
}
\details{
The fn parameter is used to generate drawdowns either from raw drawdown data (s) or from the gfamodel clas object model. When NULL, raw original values of s are used. Available keywords are:
\itemize{
    \item "birsoy", the Birsoy and Summers (1980) time is calculated based on the q series of pumping rates. q is there a a double columns matrix with column 1 giving the time (since the begining of the test) at which the period ends, and column 2 giving the flow rate during the period.
    \item "agarwal", result is the equivalent Agarwal (1980) time for recovery tests.
          Agarwal has shown in 1980 that recovery test can be interpreted with the same solutions than pumping test if one interprets the residual drawdown sr = s(t) - s(end of pumping) as a function of an equivalent time that is computed from the pumping rates history. The theory is based on the superposition principle.
          t should be the time since the begining of the recovery period. q can be either a unique value for the duration of the pump test (constant pumping rate test) or a double columns matrix with column 1 giving the time (since the begining of the test) at which the period ends, and column 2 giving the flow rate during the period (variable pumping rates).
    \item "variablerate", the drawdowns are calculated for the specified dates t, based on the model parameters (aquifer parameters aqpar should be given as well as test data testdata with pumping rate Q=1) and on the pumping rate data q. q is there a double columns matrix with column 1 giving the time (since the begining of the test) at which the period ends, and column 2 giving the flow rate during the period (variable pumping rates).
    }

The function to be used for the computation of the derivative points can be set in the miscDdOptions parameters list, with the defunc keyword. The default value is the spilne function. Available options are:
\itemize{
    \item "spline", this function approximates logarithmic derivative with spline, usign the smooth.spline function of the stats package. Use df keyword to set the degree of freedom (see \code{\link{smooth.spline}}). Default value for df is 2/3 of the number of data points.
    \item "std", this function approximates logarithmic derivative with centered differences.
    \item "bourdet", this function approximates logarithmic derivative with Bourdet's formula. Use df keyword to set the distance parameter of the Bourdet function. Default value is 2.
    \item "horne", this function approximates logarithmic derivative with the Horne's formula. pproximate logarithmic derivative with Bourdet's formula
    }
}
\examples{
# Generate drawdowns object for a standard series of t and s
MyDd <- drawdowns(t = pumptest$ths_ds1[,1], s = pumptest$ths_ds1[,2])
MyDd <- drawdowns(t = pumptest$ths_ds1[,1:2])

# Generate drawdowns object for a series of date knowing
# aquifer parameters (T = 1E-3 m2/s and S = 5E-4) and
# pumping conditions (pumping rate 3.6E-1 m3/s and distance 50 m)
MyDd <- drawdowns(t = (10^seq(2,5,length=30)),
           model = gfamodel("ths", aqpar = c(1E-3, 5E-4),
           testdata = c(3.6E-1, 50)))

# Generate drawdowns object for a large series of t and s after sampling
MyDd <- drawdowns(pumptest$thc_ds2[,1:2], sample = TRUE,
           miscDdOptions = list(nval = 70))

# Generate drawdowns object from recovery data for a constante rate pumping test
MyDd <- drawdowns(pumptest$agt_ds1, fn = "agarwal", q = 240*60,
          miscDdOptions = list(df = 10, defunc="spline"))

# Generate normalised drawdowns object from variable rate pumping test
q<- cbind(c(1800, 4800, 7800),c(500/24/60/60, 700/24/60/60, 600/24/60/60))
MyDd <- drawdowns(pumptest$tmr_ds1, fn="birsoy", q=q)
}
\references{
Renard, Philippe (2017). Hytool: an open source matlab toolbox for the interpretation of hydraulic tests using analytical solutions. Journal of Open Source Software, 2(19), 441, \doi{10.21105/joss.00441}.

Agarwal, R.G., 1980. A new method to account for producing time effects when drawdown type curves are used to analyse pressure buildup and other test data. Proceedings of the 55th Annual Fall Technical Conference and Exhibition of the Society of Petroleum Engineers. Paper SPE 9289, \doi{10.2118/9289-MS}

Birsoy YK, Summers WK (1980) Determination of aquifer parameters from step tests and intermittent pumping data. Ground Water 18(2):137–146, \doi{10.1111/j.1745-6584.1980.tb03382.x}.
}
\seealso{
\code{\link{smooth.spline}},  \code{\link{gfamodel}}, \code{\link{robreg.filter}}
}
