% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfamodel.R
\name{fit.gfamodel}
\alias{fit.gfamodel}
\title{Fit parameters of the analytical solution}
\usage{
fit.gfamodel(x, dd)
}
\arguments{
\item{x}{A gfamodel object. See details.}

\item{dd}{A drawdowns object used for fitting the analytical parameters of the returned gfamodel object.}
}
\value{
An object of class gfamodel of the same model type as the \verb{x} object in argument. Analytical solution parameters \verb{aspar} are adjusted to fit the \verb{dd} object in argument. If test data parameters \verb{testdata} are given in the \verb{x} object, it is copied in the return object and the aquifer parameters \verb{aqpar} are computed.
}
\description{
Function that fit the parameters of the analytical solution of a gfamodel object to the drawdowns describes in a drawdowns object.
}
\details{
The analytical solution parameters of the \verb{x} object are used as initial parameters. The general-purpose optimization function \code{\link{optimx}} is used. Result from the best fitting from 3 methods ("Nelder-Mead","nlminb", "L-BFGS-B") is used.
}
\seealso{
\code{\link{optimx}}
}
