% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfamodel-class.R, R/show.R
\docType{class}
\name{gfamodel-class}
\alias{gfamodel-class}
\alias{show,gfamodel-method}
\title{Class "gfamodel"}
\usage{
\S4method{show}{gfamodel}(object)
}
\arguments{
\item{object}{A gfamodel class object}
}
\description{
This class is used to reprent the groundwater flow analytical models.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Display the object, by printing. Print everay parameters of the gfamodel object, starting with the applicable analytical model type. Also print the analytical model parameters, the aquifer parameters and the test data. When parameters are not yet defined, prints 0 value.
}}

\section{Slots}{

 \describe{
   \item{\code{model}:}{Matrix of class \code{"character"}, containing the keyword reprensenting the analytical model.}
   \item{\code{description}:}{Object of class \code{"character"}, containing the detailed description of the analytical model.}
   \item{\code{aspar}:}{Object of class \code{"data.frame"}, containing the list of the analytical parameters and the corresponding numerical values (when defined).}
   \item{\code{aqpar}:}{Object of class \code{"data.frame"}, containing the list of the aquifer parameters and the corresponding numerical values (when defined).}
   \item{\code{testdata}:}{Object of class \code{"data.frame"}, containing the list of the test data and the corresponding numerical values (when defined).}
 }
}

\author{
François Bertone
}
