% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfamodel.R
\name{gfamodel}
\alias{gfamodel}
\title{Create gfamodel objects}
\usage{
gfamodel(model, aspar = NULL, aqpar = NULL, testdata = NULL)
}
\arguments{
\item{model}{A keyword or a gfamodel object. When a keyword is given, it should be one of a known model from the following list: "ths", "thc", "thn", "htj", "blt", "war" or "grg". See details}

\item{aspar}{Parameters for the analytical solution for model computation. It can be obtained from a guess fuction, adjusted manually, or obtained from a fit function}

\item{aqpar}{Aquifer's parameters for the selected model (such as Transmissivity and Storage coefficient for the Theis model). See details}

\item{testdata}{If provided, information about pumping conditions used for calculating aquifer parameters \verb{aqpar} from analytical solution parameters \verb{aspar} or analytical solution parameters \verb{aspar} from aquifer parameters \verb{aqpar}}
}
\value{
An object of class gfamodel the type of selected \verb{model}, the corresponding description, and model parameters of type \verb{aqpar} or \verb{aspar}.
    If \verb{model} is a gfamodel object, the returned gfamodel object is constructed from this object and in priority order either the analytical solution parameters (if avilable) or the aquifer parameters. A \verb{testsdata} parameter should then be provided.
}
\description{
Create objects of class gfamodel, i.e. objects that hold the information relative to a known groundwater flow analytical model
}
\details{
If model is set from keywords, the options to date are:
\itemize{
    \item "ths" for a Theis (1935) solution. Analytical solution parameters are: \verb{a}, the slope of Jacob Straight Line; and \verb{t0}, intercept with the horizontal axis for s = 0. When the model is constructed from aquifer parameters and test data, \verb{aqpar} should be a vector of 2 numeric (Transmissivity and Storativity), and \verb{testdata} a vector of 2 numerics (Pumping rate and Distance to the pumping well).
    \item "thc" for a confined aquifer with a boundary effect using the Theis (1941) solution. Analytical solution parameters are: \verb{a}, the slope of Jacob Straight Line; \verb{t0}, intercept of the first segment of straight line; and \verb{ti} the time of intersection between the 2 straight lines. When the model is constructed from aquifer parameters and test data, \verb{aqpar} should be a vector of 3 numeric (Transmissivity, Storativity, Distance to image well), and \verb{testdata} a vector of 2 numerics (Pumping rate, Distance to the pumping well).
    \item "thn" for a confined aquifer with an impermeable boundary with the Theis solution. Analytical solution parameters are: \verb{a}, the slope of Jacob Straight Line; \verb{t0}, intercept of the first segment of straight line; and \verb{ti} the time of intersection between the 2 straight lines. When the model is constructed from aquifer parameters and test data, \verb{aqpar} should be a vector of 3 numeric (Transmissivity, Storativity, Distance to image well), and \verb{testdata} a vector of 2 numerics (Pumping rate, Distance to the pumping well).
    \item "htj" for a confined aquifer with leakage with the Hantush and Jacob (1955) model. Analytical solution parameters are: \verb{a}, the slope of Jacob Straight Line; and \verb{t0}, intercept with the horizontal axis for s = 0; and \verb{r/B}. When the model is constructed from aquifer parameters and test data, \verb{aqpar} should be a vector of 3 numeric (Transmissivity, Storativity, Aquitard conductivity), and \verb{testdata} a vector of 3 numerics (Pumping rate, Distance to the pumping well, Thickness of the aquitard).
    \item "blt" for a Boulton (1963) model for unconfined aquifer. Analytical solution parameters are: slope\verb{a} of the late time Jacob's straight line"; intercept time \verb{t0} with the horizontal axis for s = 0; intercept \verb{t1}; and empirical parameter \verb{phi} that trigger the delay. When the model is constructed from aquifer parameters and test data, \verb{aqpar} should be a vector of 3 numeric (Transmissivity, Storativity, Drainage Porosity), and \verb{testdata} a vector of 2 numerics (Pumping rate and Distance to the pumping well).
    \item "grg" for a Gringarten and Ramey (1974) model for infinite conductivity fracture in a confined aquifer. Analytical solution parameters are slope \verb{a} and intercept \verb{t0}. When the model is constructed from aquifer parameters and test data, \verb{aqpar} should be a vector of 2 numeric (Transmissivity and Sxf2), and \verb{testdata} a numeric (Pumping rate)
    \item "war" for the Warren and Root (1965) solution. Analytical solution parameters are: the slope \verb{a} of the Jacob straight line, the time \verb{t0} when the horizontal axis intercept with the early time asymptote, the time \verb{t1} when the horizontal axis intercept with the late time asymptote, the time \verb{tm} of the minimum of the derivative. When the model is constructed from aquifer parameters and test data, \verb{aqpar} should be a vector of 4 numeric (Transmissivity, Storativity Sf, Storativity Sm, Interporosity flow lamda), and \verb{testdata} a vector of 2 numerics (Pumping rate and Distance to the pumping well).
    \item "grf" for the Barker (1988) general radial flow model. Analytical solution parameters are: \verb{a} equivalent of the slope of Jacob model, \verb{t0} equivalent of the time of Jacob model, \verb{n} the fractional flow dimension. When the model is constructed from aquifer parameters, \verb{aqpar} should be a vector of 3 numeric (equivalent transmissivity Tn, equivalent storativity Sn, flow dimension n). The \verb{testdata} parameters are compulsory for the construction of the gfamodel object, as a vector of 3 numerics (pumping rate Q, distance r to the pumping well, radius rw of the pumping well).
    }

If \verb{model} is an existing gfamodel, \verb{testsdata} parameters should be given: the \verb{aspar} of the resulting object is then calculated if \verb{model} have valid \verb{aspar}, or the \verb{aqpar} of the resulting model is calculated if model have valid \verb{aspar}.
}
\examples{
# Load samples test data
data("pumptest")
# Construct initial guess of a Hantush and Jacob model for the htj_ds1 data set
mod1 <- guess.gfamodel(gfamodel("htj"), drawdowns(pumptest$htj_ds1))
# If the first guess is not satisfactory, adjust the model manually
mod1 <- gfamodel("htj", aspar = c(5.3,0.3,4e5))
# perform automatic fitting and plot result
mod2 <- fit.gfamodel(mod1, dd = drawdowns(pumptest$htj_ds1))
plot(pumptest$htj_ds1, model = mod2)
print(mod2)
}
\references{
Renard, Philippe (2017). Hytool: an open source matlab toolbox for the interpretation of hydraulic tests using analytical solutions. Journal of Open Source Software, 2(19), 441, \href{http://joss.theoj.org/papers/10.21105/joss.00441}{doi:10.21105/joss.00441}
}
\seealso{
\code{\link{fit.gfamodel}}, \code{\link{guess.gfamodel}}
}
