% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfamodel.R
\name{guess.gfamodel}
\alias{guess.gfamodel}
\title{First guess for the analytical solution parameters}
\usage{
guess.gfamodel(x, dd)
}
\arguments{
\item{x}{A gfamodel object. See details.}

\item{dd}{An drawdowns object used for extracting a first guess.}
}
\value{
An object of class gfamodel of the same type than the entry \verb{x} object with new analytical solution parameters. If available in the entry \verb{x} object, aquifer parameters and test data are copied in the constructed object.
}
\description{
Function that guess intial analytical solution parameters for fitting a gfamodel to a drawdowns object
}
\details{
Only model type of the entry \verb{x} object is used. If available, in the analytical solution parameters of the \verb{x} object are ignored.
}
