% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,drawdowns-method}
\alias{plot,drawdowns-method}
\title{Class "drawdowns"}
\usage{
\S4method{plot}{drawdowns}(x, model = NULL, log = "xy", col = c("blue4",
  "aquamarine4", "deepskyblue3", "darkolivegreen3"), derivative = TRUE,
  legend = TRUE, leg.pos = "bottomright", unit.slope = FALSE,
  ylab = "Drawdown", xlab = "Time")
}
\arguments{
\item{x}{drawdowns object to be plotted}

\item{model}{gfamodel object to be superposed to \verb{x}. Data are calculated for the \verb{x} time series. if NULL no model is plotted.}

\item{log}{character string which contains "x" if the x axis is to be logarithmic, "y" if the y axis is to be logarithmic and "xy" or "yx" if both axes are to be logarithmic.}

\item{col}{colors vector for the x data, model data, x derivative and model derivative respectively. If there are fewer colors than 4, color are recycled.}

\item{derivative}{logical. If TRUE the derivatives for \verb{x} argument, and for \verb{model} argument if any, are ploted.}

\item{legend}{logical. If TRUE a lengend is printed.}

\item{leg.pos}{single keyword setting the location of the legend, from the list "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center".}
}
\description{
Function that plots drawdowns objects.
}
\details{
Plot drawdowns and derivative of drawdown curves for the \verb{x} object. If a the \verb{model} argument is set, supperpose a line for the corresponding drawdowns and derivative.
}
