% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,gfamodel-method}
\alias{plot,gfamodel-method}
\title{Class "gfamodel"}
\usage{
\S4method{plot}{gfamodel}(x, title = TRUE, legend = TRUE, log = "xy")
}
\arguments{
\item{x}{gfamodel object to be plotted}

\item{title}{logical. Indicates if the analytical model type is to be printed}

\item{legend}{logical. Indicates if the legend type is to be printed}

\item{log}{character string which contains "x" if the x axis is to be logarithmic, "y" if the y axis is to be logarithmic and "xy" or "yx" if both axes are to be logarithmic.}
}
\description{
Function that plots gfamodel objects.
}
\details{
Plot standard drawdowns and derivative of drawdown curves for the analytical model set in the \verb{x} object.
}
